+++ src/nova/trame/model/ornl/oncat_data_selector.py
@@ -1,6 +1,7 @@
 """ONCat backend for NeutronDataSelector."""
 
 import os
+import re
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
@@ -134,8 +135,11 @@ class ONCatDataSelectorModel(NeutronDataSelectorModel):
             else:
                 can_add = True
 
-            if self.state.search and self.state.search.lower() not in os.path.basename(path).lower():
-                can_add = False
+            if can_add and self.state.show_filter and self.state.filter:
+                if self.state.use_regex == [0]:
+                    can_add = bool(re.search(rf"{self.state.filter}", os.path.basename(path)))
+                else:
+                    can_add = self.state.filter.lower() in os.path.basename(path).lower()
 
             if can_add:
                 datafiles.append(self.create_datafile_obj(datafile_data, projection))